+++ __init__.py
@@ -38,7 +38,9 @@ PROPS = [
     ('sphere_scale', bpy.props.FloatVectorProperty(name='Scale', description='Scale of the training sphere in xyz axes', default=(1.0, 1.0, 1.0), update=helper.properties_ui_upd) ),
     ('sphere_radius', bpy.props.FloatProperty(name='Radius', description='Radius scale of the training sphere', default=4.0, soft_min=0.01, unit='LENGTH', update=helper.properties_ui_upd) ),
     ('seed', bpy.props.IntProperty(name='Seed', description='Random seed for sampling views on the training sphere', default=0) ),
-    ('nb_cameras', bpy.props.IntProperty(name='Cameras', description='Number of fixed cameras placed on the sphere', default=100, soft_min=1) ),
+    ('nb_cameras', bpy.props.IntProperty(name='Cameras', description='Number of fixed cameras placed on the sphere', default=30, soft_min=1) ),
+    ('first_frame_nr', bpy.props.IntProperty(name='Start Frame Number', description='First frame of the animation to render', default=1, soft_min=1) ),
+    ('final_frame_nr', bpy.props.IntProperty(name='End Frame Number', description='Last frame of the animation to render', default=48, soft_min=1) ),
     ('show_sphere', bpy.props.BoolProperty(name='Preview Sphere', description='Whether to show the training sphere from which random views will be sampled', default=False, update=helper.visualize_sphere) ),
     ('upper_views', bpy.props.BoolProperty(name='Upper Views', description='Whether to sample views from the upper hemisphere of the training sphere only', default=False) ),
     ('outwards', bpy.props.BoolProperty(name='Outwards', description='Whether to point the camera outwards of the training sphere', default=False, update=helper.properties_ui_upd) ),
