+++ render_operator.py
@@ -16,6 +16,9 @@ class RenderScene(bpy.types.Operator):
     def execute(self, context):
         scene = context.scene
         # clean directory name (unsupported characters replaced) and output path
+        #TODO: Make sure all meta data, point clouds are written to file at the beginning of rendering
+        #TODO: Make sure the user-specified end frame number is set correctly just before rendering, make sure rendering stops at the correct frame
+
         output_dir = bpy.path.clean_name(scene.dataset_name)
         output_path = os.path.join(scene.save_path, output_dir)
         os.makedirs(output_path, exist_ok=True)
